refactor: move credential card status
